requirements.txt
# Web Framework
django>=5.0.2
channels>=4.0.0
channels-redis==4.1.0

# Database
psycopg2-binary>=2.9.9
asyncpg==0.29.0
dj-database-url>=2.1.0

# Web Scraping
selenium==4.17.2
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-django==4.8.0
pytest-cov==4.1.0
locust==2.21.0

# Monitoring
sentry-sdk>=1.40.0
prometheus-client==0.19.0

# Development
python-dotenv>=1.0.0
black==24.1.1
flake8==7.0.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
uvicorn==0.27.1 